[cli] Restructure and clean up init command
