+++ packages/@sanity/cli/src/actions/init-plugin/initPlugin.js
