+++ packages/@sanity/cli/src/actions/init-plugin/pluginTemplates.js
