+++ packages/@sanity/cli/src/actions/init-project/bootstrapTemplate.js
