+++ packages/@sanity/cli/src/actions/init-project/createManifest.js
