+++ packages/@sanity/cli/src/actions/init-project/promptForDatasetName.js
