+++ packages/@sanity/cli/src/actions/init-project/templates/clean.js
