+++ packages/@sanity/cli/src/actions/init-project/templates/index.js
