+++ packages/@sanity/cli/src/actions/init-project/templates/moviedb.js
