+++ packages/@sanity/util/src/index.js
@@ -1,6 +1,7 @@
 import getConfig from './getConfig'
 import reduceConfig from './reduceConfig'
 import lazyRequire from './lazyRequire'
+import dynamicRequire from './dynamicRequire'
 import safeJson from './safeJson'
 import getSanityVersions from './getSanityVersions'
 import * as pathTools from './pathTools'
@@ -9,6 +10,7 @@ export {
   getConfig,
   reduceConfig,
   lazyRequire,
+  dynamicRequire,
   safeJson,
   getSanityVersions,
   pathTools
